feat: Add CSS, remove html img widths
Added some simple CSS for practice. I was having some trouble centering
the ordered and unordered lists. So I removed the list feature on the
index file. And I made do with padding for the recipe pages.

Once I learn more about placing things on the page with CSS, I should
probably update the `.ingredients` and `.steps` styles to the better
method.